fix: added Suspense for useSearchParams error  fix
